ws2812: Enable timer clock prior to timer init
Turn on the timer clock before configuring the timer. We
were inadvertently depending on auto-generated HAL init
code for turning on the timer clock, even though we were
already explicitly turning this clock on after init.

So, just turn on the timer clock before configuring the
timer, which removes our last dependency on the HAL init
code. With this, we are able to use TIM3 or TIM4 for LED
refresh, without needing to configure them in CubeMX.

Signed-off-by: evilwombat <evilwombat@peregrine.local>